logsentinel/requirements.txt
pandas>=2.0
python-dateutil>=2.8
watchdog>=3.0
scikit-learn>=1.2
PyYAML>=6.0
